+++ internal/home/config.go
@@ -179,6 +179,8 @@ var config = configuration{
 			RefuseAny:          true,
 			AllServers:         false,
 
+			TrustedProxies: []string{"127.0.0.0/8", "::1/128"},
+
 			// set default maximum concurrent queries to 300
 			// we introduced a default limit due to this:
 			// https://github.com/AdguardTeam/AdGuardHome/issues/2015#issuecomment-674041912